pyproject.toml
[project]
name = "a-view"
version = "0.0.1"
description = "LibreOffice 기반 문서 Viewer "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "uvicorn[standard]>=0.35.0",
]

[project.scripts]
a-view = "a_view.main:main"
